+++ app/src/main/AndroidManifest.xml
@@ -113,7 +113,7 @@
             android:name=".presentation.mypage.terms.WebViewActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.login.IntroActivity"
+                android:name=".presentation.intro.IntroActivity"
             android:exported="true">
             <!-- 기본 런처 필터 -->
             <intent-filter>
