+++ app/src/main/java/com/eatssu/android/presentation/intro/IntroActivity.kt
@@ -1,17 +1,18 @@
-package com.eatssu.android.presentation.login
+package com.eatssu.android.presentation.intro
 
 import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
 import androidx.activity.viewModels
-import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.lifecycleScope
 import com.eatssu.android.databinding.ActivityIntroBinding
 import com.eatssu.android.presentation.MainActivity
 import com.eatssu.android.presentation.UiEvent
 import com.eatssu.android.presentation.UiState
+import com.eatssu.android.presentation.base.BaseActivity
+import com.eatssu.android.presentation.login.LoginActivity
 import com.eatssu.android.presentation.util.showToast
 import com.eatssu.android.presentation.util.startActivity
 import com.eatssu.common.EventLogger
+import com.eatssu.common.enums.FirebaseScreenId
 import com.eatssu.common.enums.LaunchPath
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collectLatest
@@ -21,13 +22,9 @@ import kotlinx.coroutines.launch
 class IntroActivity : AppCompatActivity() {
 
     private val introViewModel: IntroViewModel by viewModels()
-    private lateinit var binding: ActivityIntroBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = ActivityIntroBinding.inflate(layoutInflater)
-        enableEdgeToEdge()
-        setContentView(binding.root)
         log()
 
         lifecycleScope.launch {
