+++ app/src/main/res/layout/activity_intro.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/primary"
-    tools:context=".presentation.login.IntroActivity">
+        tools:context=".presentation.intro.IntroActivity">
 
     <ImageView
         android:layout_width="250dp"
